requirements.txt
absl-py==2.2.2
astunparse==1.6.3
beautifulsoup4==4.13.3
bleach==6.2.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
datatable==1.1.0
et_xmlfile==2.0.0
filelock==3.17.0
flatbuffers==25.2.10
fonttools==4.56.0
gast==0.4.0
gdown==5.2.0
google-auth==2.40.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
imbalanced-learn
inflate64==1.0.1
joblib==1.4.2
kaggle==1.6.17
keras==2.13.1
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib
mdurl==0.1.2
ml-dtypes==0.3.2
multivolumefile==0.2.3
namex==0.0.9
numpy
oauthlib==3.2.2
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas
pillow==11.1.0
protobuf==4.25.7
psutil==7.0.0
py7zr==0.22.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybcj==1.0.3
pycryptodomex==3.21.0
Pygments==2.19.1
pyparsing==3.2.1
pyppmd==1.1.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.1
pyzstd==0.16.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rs-datasets==0.5.1
rsa==4.9.1
scikit-learn
scipy==1.15.2
seaborn
six==1.17.0
sklearn-compat==0.1.3
soupsieve==2.6
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.13.0
tensorflow-metal==0.7.0
termcolor==3.1.0
text-unidecode==1.3
texttable==1.7.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.5.0
tzdata==2025.1
urllib3==2.3.0
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.2
xgboost
xlrd==2.0.1
